requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
pandas>=2.0.0
joblib>=1.3.0
openpyxl
